A draft of keys (aka normalizedId) working; see details;
Well, this is a lot to take, I discovered what I needed to do mostly in reverse order, but I will explain in order to make better sense.

1. The concept of `cwd` was broken. We needed more meaninful name. In Mendel v1 we used basedir for that, and browserify, webpack, node-resolver, browser-resolver, all use this concept. process.cwd() is a commandline concept and just a fallback to projects without .mendelrc, so in order to fix that I needed to trace all places `cwd` was ambiguous. I decided to change the name to `projectRoot`, with the intention of people onboarding on the project to never mistake the concept again.

2. There was ambiguity about variations, and there might still be places to track and replace. Because there is difference between a list of variational folders, a list of variations to process on a per file bases, and a "variation chain", I reverted back to stable code from mendel v1. Than, I found all places that mendel pipeline or it's sub-dependencies were calling "chain" that didn't refer to a chain. For reference:

WIP: this is still ambiguous and need work.

variations: an array of variation objects
variation: an object with id and chain array
chain: an array of paths relative to projectRoot
allDirs: an array of unique variation paths relative to projectRoot
lookupChains: an array of modified-chains (without basetree) to resolve multi-layer scenarios

former ambiguity:
basedir -> projectRoot : project root (but since we use "base" as the "default" or "master" variation, this makes hard to use this name)
basetree -> baseConfig.dir : the folder relative to projectRoot that contains the base variation
variationdir could refer to the parent folder of all variations in mendel v1 or one item in a chain
variationroot old name for variationdir when meaning the folder containing variation folders

I am tired right now, not making sense anymore, will rewrite this and force push later

@stephanwlee